refactor(lib): collapse the four duplicated single sources
G7 of audit-refactor-7.

escapeHtml had two copies; the content one also escaped `'`. That gets
dropped rather than widened — overlay has no single-quoted attributes,
and widening would change every apostrophe in the clipboard's text/html,
logs.html, Asana html_notes and the live preview, with nothing to catch
it. The decision is pinned by a test so the next batch does not redo the
arithmetic.

imageCell existed three times and, unlike its sibling functions, none of
them escaped the link text. It moves to issueBodyShared, and
escapeMdLinkText has to move with it — calling it from the leaf while it
lived in buildIssueMarkdown would close a cycle.

MarkdownIt was configured in four places, each re-enabling strikethrough
on the next line; the factory keeps per-file instances so one caller's
md.use() cannot leak into the others.

Both are locked by definition-count scans rather than name greps.

ActionLogContent's frozen-phase set and its local sky/red-700 literals
converge too. The tone colours genuinely change (sky→blue, red-700→
red-600) because the local values had drifted from TONE_TEXT; contrast
improves for the tag and drops for the value, and prd records both
numbers — matching the app-wide table beats keeping this one spot better.
